requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Object Detection & Tracking
ultralytics>=8.0.0  # YOLOv8/YOLOv11
supervision>=0.16.0  # Tracking utilities and annotations

# Image Processing
scikit-image>=0.21.0
scipy>=1.11.0
Pillow>=10.0.0

# Video Processing
ffmpeg-python>=0.2.0

# Deep Learning Utils
onnxruntime>=1.15.0  # For optimized inference

# CLI and Configuration
click>=8.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: GUI
gradio>=4.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
